@@ -6,43 +6,38 @@
   "license": "BSD-3-Clause",
   "type": "module",
   "scripts": {
-    "dev": "npm run build:local && yext pages generate-test-data && pages dev",
-    "build:local": "vite build",
-    "build": "yext pages build && yext pages generate-test-data && yext pages render",
-    "build:serve": "npm run build && yext pages serve",
-    "serve": "yext pages serve",
-    "features": "pages generate features",
-    "fmt": "prettier -w .",
-    "lint": "eslint .",
-    "clean": "rm -rf .artifact-output .yext/ dist/ localData sites-rendered-output",
-    "generate-notices": "generate-license-file --input package.json --output ./THIRD-PARTY-NOTICES --overwrite"
+    "dev": "pages dev",
+    "prod": "pages prod",
+    "build": "pages build"
   },
   "engines": {
-    "node": ">=17"
+    "node": "^18.0.0 || >=20.2.0"
   },
   "dependencies": {
-    "@yext/react-components": "^0.0.5",
-    "clsx": "^1.2.1",
+    "@headlessui/react": "^1.7.12",
+    "@heroicons/react": "^2.0.14",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "classnames": "^2.4.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "tailwind-merge": "^1.6.2"
+    "react-phone-number-input": "^3.2.19",
+    "tailwind-merge": "^2.2.0"
   },
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.9",
     "@types/react": "^17.0.49",
     "@types/react-dom": "^17.0.17",
-    "@typescript-eslint/eslint-plugin": "^5.36.2",
-    "@typescript-eslint/parser": "^5.36.2",
-    "@vitejs/plugin-react": "^2.1.0",
-    "@yext/pages": "1.0.0-beta.22",
+    "@typescript-eslint/eslint-plugin": "^6.16.0",
+    "@typescript-eslint/parser": "^6.16.0",
+    "@vitejs/plugin-react": "^4.2.1",
+    "@yext/pages": "1.0.0",
+    "@yext/pages-components": "^1.0.0-rc.1",
     "autoprefixer": "^10.4.8",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-react": "^7.31.7",
-    "generate-license-file": "^1.3.0",
-    "postcss": "^8.4.16",
-    "prettier": "^2.7.1",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.8.3",
-    "vite": "^3.1.0"
+    "postcss": "^8.4.32",
+    "prettier": "^3.1.1",
+    "tailwindcss": "^3.4.0",
+    "typescript": "^5.3.3",
+    "vite": "^5.1.4"
   }
 }
