@@ -1,4 +1,6 @@
-module.exports = {
+import type { Config } from "tailwindcss";
+
+export default {
   content: ["./src/**/*.{html,js,jsx,ts,tsx}"],
   theme: {
     extend: {
@@ -6,7 +8,10 @@ module.exports = {
         orange: "#ff9500",
         "dark-orange": "#db8000",
       },
+      scale: {
+        1.02: "1.02",
+      },
     },
   },
-  plugins: [require("@tailwindcss/typography")],
-};
+  plugins: [],
+} satisfies Config;
